testsuite: preparation to avoid code duplication in cxx
mtest.cxx differs from its C equivallent enough to warrant a separate
framework. This result in many code duplication. Every time we
add/modify features in mtest, we need duplicate code in both.

This commit adds `mtest_common.c` that will get linked to both c and cxx
tests, and to be used for common utility code shared between c and cxx
framework.
