+++ test/mpi/cxx/Makefile.am
@@ -9,6 +9,6 @@ include $(top_srcdir)/Makefile_cxx.mtest
 
 EXTRA_DIST = testlist.in
 
-static_subdirs = util attr pt2pt init info comm coll datatype errhan topo
+static_subdirs = attr pt2pt init info comm coll datatype errhan topo
 SUBDIRS      = $(static_subdirs) $(iodir) $(rmadir) $(spawndir)
 DIST_SUBDIRS = $(static_subdirs) io rma spawn
