+++ test/mpi/include/mpitest.h
@@ -11,6 +11,7 @@
 #include <mpi.h>
 #include "mpitestconf.h"
 #include "mpithreadtest.h"
+#include "mtest_common.h"
 
 /*
  * Init and finalize test
@@ -29,13 +30,6 @@ int MTestReturnValue(int);
  */
 void MTestSleep(int);
 void MTestGetDbgInfo(int *dbgflag, int *verbose);
-
-typedef void MTestArgList;
-MTestArgList *MTestArgListCreate(int argc, char *argv[]);
-char *MTestArgListGetString(MTestArgList * head, const char *arg);
-int MTestArgListGetInt(MTestArgList * head, const char *arg);
-long MTestArgListGetLong(MTestArgList * head, const char *arg);
-void MTestArgListDestroy(MTestArgList * head);
 
 int MTestGetIntracomm(MPI_Comm *, int);
 int MTestGetIntracommGeneral(MPI_Comm *, int, int);
