+++ test/mpi/include/mpitestcxx.h
@@ -8,8 +8,20 @@
 #ifndef MPITESTCXX_H_INCLUDED
 #define MPITESTCXX_H_INCLUDED
 
-#ifndef MPITESTCONF_H_INCLUDED
-#error Required mpitestconf.h file not included first!
+#include "mpi.h"
+#include "mpitestconf.h"
+
+extern "C" {
+#include "mtest_common.h"
+}
+
+#ifdef HAVE_IOSTREAM
+#include <iostream>
+#ifdef HAVE_NAMESPACE_STD
+using namespace std;
+#endif
+#else
+#include <iostream.h>
 #endif
 
 #include <string.h>
@@ -22,13 +34,6 @@ void MTestPrintError(int);
 void MTestPrintErrorMsg(const char[], int);
 void MTestPrintfMsg(int, const char[], ...);
 void MTestError(const char[]);
-
-typedef void MTestArgList;
-MTestArgList *MTestArgListCreate(int argc, char *argv[]);
-char *MTestArgListGetString(MTestArgList * head, const char *arg);
-int MTestArgListGetInt(MTestArgList * head, const char *arg);
-long MTestArgListGetLong(MTestArgList * head, const char *arg);
-void MTestArgListDestroy(MTestArgList * head);
 
 int MTestGetIntracomm(MPI::Intracomm &, int);
 int MTestGetIntracommGeneral(MPI::Intracomm &, int, bool);
