+++ test/mpi/util/mtest_cxx.cxx
@@ -4,22 +4,7 @@
  *  (C) 2001 by Argonne National Laboratory.
  *      See COPYRIGHT in top-level directory.
  */
-#include "mpi.h"
-#include "mpitestconf.h"
-#ifdef HAVE_IOSTREAM
-// Not all C++ compilers have iostream instead of iostream.h
-#include <iostream>
-#ifdef HAVE_NAMESPACE_STD
-// Those that do often need the std namespace; otherwise, a bare "cout"
-// is likely to fail to compile
-using namespace std;
-#endif
-#else
-#include <iostream.h>
-#endif
 #include "mpitestcxx.h"
-#include <stdlib.h>
-#include <string.h>
 
 static int dbgflag = 0;         /* Flag used for debugging */
 static int wrank = -1;          /* World rank */
@@ -569,121 +554,3 @@ static void MTestRMACleanup(void)
 {
 }
 #endif
-
-/* ------------------------------------------------------------------------ */
-/* All dtpools related code */
-
-typedef struct MTestArgListEntry {
-    char *arg;
-    char *val;
-    struct MTestArgListEntry *next;
-} MTestArgListEntry;
-
-static char *progname;
-
-static void MTestArgListInsert(MTestArgListEntry ** head, char *arg, char *val)
-{
-    MTestArgListEntry *tmp = *head;
-
-    if (!tmp) {
-        tmp = (MTestArgListEntry *) malloc(sizeof(MTestArgListEntry));
-        tmp->arg = arg;
-        tmp->val = val;
-        tmp->next = NULL;
-        *head = tmp;
-        return;
-    }
-
-    while (tmp->next)
-        tmp = tmp->next;
-
-    tmp->next = (MTestArgListEntry *) malloc(sizeof(MTestArgListEntry));
-    tmp->next->arg = arg;
-    tmp->next->val = val;
-    tmp->next->next = NULL;
-}
-
-static char *MTestArgListSearch(MTestArgListEntry * head, const char *arg)
-{
-    char *val = NULL;
-
-    while (head && strcmp(head->arg, arg))
-        head = head->next;
-
-    if (head)
-        val = head->val;
-
-    return val;
-}
-
-static void MTestArgListPrintError(const char *arg)
-{
-    fprintf(stdout, "Error: argument -%s= has not been defined!\n", arg);
-    exit(-1);
-}
-
-void MTestArgListDestroy(MTestArgList * head)
-{
-    MTestArgListEntry *cur = (MTestArgListEntry *) head;
-
-    while (cur) {
-        MTestArgListEntry *prev = cur;
-        cur = cur->next;
-        free(prev->arg);
-        free(prev->val);
-        free(prev);
-    }
-
-    free(progname);
-}
-
-/*
- * following args are expected to be of the form: -arg=val
- */
-MTestArgList *MTestArgListCreate(int argc, char *argv[])
-{
-    int i;
-    char *string = NULL;
-    char *tmp = NULL;
-    char *arg = NULL;
-    char *val = NULL;
-
-    MTestArgListEntry *head = NULL;
-
-    for (i = 1; i < argc; i++) {
-        /* extract arg and val */
-        string = strdup(argv[i]);
-        tmp = strtok(string, "=");
-        arg = strdup(tmp + 1);  /* skip prepending '-' */
-        tmp = strtok(NULL, "=");
-        val = strdup(tmp);
-
-        MTestArgListInsert(&head, arg, val);
-
-        free(string);
-    }
-
-    progname = strdup(argv[0]);
-
-    return head;
-}
-
-char *MTestArgListGetString(MTestArgList * head, const char *arg)
-{
-    char *tmp;
-
-    if (!(tmp = MTestArgListSearch((MTestArgListEntry *) head, arg)))
-        MTestArgListPrintError(arg);
-
-    return tmp;
-}
-
-int MTestArgListGetInt(MTestArgList * head, const char *arg)
-{
-    return atoi(MTestArgListGetString(head, arg));
-}
-
-long MTestArgListGetLong(MTestArgList * head, const char *arg)
-{
-    return atol(MTestArgListGetString(head, arg));
-}
